+++ R/check_missing_transactions.R
@@ -7,7 +7,7 @@
 #' @return A data frame, filtered with the rows containing a total
 #' quantity smaller than zero.
 #' @examples
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' check_missing_transactions(formatted.ACB)
 #' @importFrom dplyr %>% rename mutate select filter bind_rows group_by slice_tail