+++ R/print_report.R
@@ -86,7 +86,7 @@
 #' @examples
 #' \donttest{
 #' list.prices <- list_prices_example
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' if (!is.null(list.prices) && !is.null(list.prices$date2)) {
 #'   print_report(formatted.ACB,