+++ R/report_overview.R
@@ -141,7 +141,7 @@
 #' losses, net, currency.
 #' @export
 #' @examples
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' report_overview(formatted.ACB, today.data = FALSE)
 #' @importFrom dplyr %>% filter mutate group_by summarize slice_tail bind_rows