+++ R/report_summary.R
@@ -104,7 +104,7 @@
 #' Type, Amount, currency.
 #' @export
 #' @examples
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' report_summary(formatted.ACB, today.data = FALSE)
 #' @importFrom dplyr %>% filter mutate group_by ungroup select summarize