+++ R/tax_box.R
@@ -123,7 +123,7 @@
 #' @export
 #' @examples
 #' my.list.prices <- list_prices_example
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' report.summary <- report_summary(formatted.ACB, today.data = TRUE, list.prices = my.list.prices)
 #' sup.losses <- get_sup_losses(formatted.ACB, 2021)