+++ R/tax_table.R
@@ -7,7 +7,7 @@
 #' @return A flextable object, with certain formatting features.
 #' @export
 #' @examples
-#' all.data <- format_shakepay(data_shakepay)
+#' all.data <- format_exchanges(data_shakepay)
 #' formatted.ACB <- format_ACB(all.data, verbose = FALSE)
 #' x <- get_sup_losses(formatted.ACB, 2021)
 #' tax_table(x)